requirements.txt
Django==5.0.7
django_recaptcha==4.0.0
django-tinymce==4.1.0

#django.contrib.admin
#django.contrib.auth
#django.contrib.contenttype
#django.contrib.session
#django.contrib.message
#django.contrib.staticfiles
django.user
#django.main
django.captcha
django-recaptcha
django-simple-captcha
django.froala-editor
django.crispy-forms
django.tinymce

#django.user.CustomUser
django.custom-user-models

#django.middleware.security.SecurityMiddleware
#django.contrib.sessions.middleware.SessionMiddleware
#django.middleware.common.CommonMiddleware
#django.middleware.csrf.CsrfViewMiddleware
#django.contrib.auth.middleware.AuthenticationMiddleware
#django.contrib.messages.middleware.MessageMiddleware
#django.middleware.clickjacking.XFrameOptionsMiddleware

